requirements.txt
# ExcelConvert 项目依赖清单
# Python 3.8+ required

# === 核心依赖 ===

# Excel 文件处理
openpyxl>=3.0.10,<4.0.0

# 数据处理
pandas>=1.3.0,<3.0.0

# YAML 配置文件解析
PyYAML>=5.4.0,<7.0.0

# LangGraph 工作流引擎
langgraph>=0.0.20,<1.0.0

# HTTP 请求（LLM API 调用）
requests>=2.25.0,<3.0.0

# 环境变量管理（.env文件支持）
python-dotenv>=0.19.0,<2.0.0

# === 可选依赖 ===

# 如果需要使用 LangChain 的其他功能
# langchain>=0.1.0

# 如果需要更好的日志输出
# colorlog>=6.7.0

# 如果需要性能分析
# profile>=1.0.0

# === 开发依赖（可选） ===

# 测试框架
# pytest>=7.0.0
# pytest-cov>=3.0.0

# 代码质量检查
# pylint>=2.12.0
# black>=21.0.0
# flake8>=3.9.0

# 类型检查
# mypy>=0.910
